move tooltip folder to /widgets
