+++ src/favorites/FavoriteButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SimpleTooltipOrigin } from '../tooltip/SimpleTooltip';
+import { SimpleTooltipOrigin } from '../widgets/tooltip/SimpleTooltip';
 import { injectIntl } from 'react-intl';
 import Icon from '../widgets/Icon';
 import './FavoriteButton.scss';
