+++ src/widgets/tooltip/ProfileTooltip.scss
