+++ src/widgets/tooltip/SimpleTooltip.js
