+++ src/widgets/tooltip/SimpleTooltip.scss
