+++ src/widgets/tooltip/tooltipHelpers.js
